Added `setAttr` module
Attach ‘data-if-set-attr` to an element which can accept a change event, and give it a set of values in the following format:

```
nameOfOutput|attrName,anotherNameOfOutput,attrName
```

As a comma separated list of items, this will separately trigger elements with the attribute `data-if-set-attr-output`, and a value which matches the string before the pipe. The value of the source element will be passed to all target elements, and inserted as an attribute matching the name after the pipe. (Camelcasing will be converted to dashcase.) This can be any standard attribute, or a custom data attribute.